+++ submission/src/epilot_submission/sdkconfiguration.py
@@ -26,9 +26,9 @@ class SDKConfiguration:
     server_idx: Optional[int] = 0
     language: str = "python"
     openapi_doc_version: str = "1.0.0"
-    sdk_version: str = "1.9.1"
-    gen_version: str = "2.446.5"
-    user_agent: str = "speakeasy-sdk/python 1.9.1 2.446.5 1.0.0 epilot-submission"
+    sdk_version: str = "1.9.2"
+    gen_version: str = "2.455.2"
+    user_agent: str = "speakeasy-sdk/python 1.9.2 2.455.2 1.0.0 epilot-submission"
     retry_config: OptionalNullable[RetryConfig] = Field(default_factory=lambda: UNSET)
     timeout_ms: Optional[int] = None