+++ submission/RELEASES.md
@@ -155,3 +155,11 @@ Based on:
 - Speakeasy CLI 1.426.0 (2.446.5) https://github.com/speakeasy-api/speakeasy
 ### Generated
 - [python v1.9.1] submission
+
+## 2024-11-09 00:25:22
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.api.epilot.io/submission-api.yaml
+- Speakeasy CLI 1.436.0 (2.455.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v1.9.2] submission